refactored StudentResultsDisplay
